Use HD key derivation from bitcoinj implementation and remove the old one

+++ src/main/java/eu/cryptoeuro/accountmapper/service/hdkey/HdAddress.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class HdAddress {
-    private Long index;
+    private int index;
     private String address;
 }
